requirements.txt
selenium==3.141.0
progressbar2==3.53.1